Added SimplePosition.java, minor renaming
