+++ core/src/com/po/fuck/model/Drawable.java
@@ -1,4 +1,4 @@
-package com.po.fuck.view;
+package com.po.fuck.model;
 
 import com.badlogic.gdx.math.Vector2;
 
@@ -6,7 +6,6 @@ public interface Drawable {
     default int get_z() {
         return 0;
     }
-
     public Vector2 getPosition();
 }
