+++ core/src/com/po/fuck/model/GameObject.java
@@ -3,7 +3,6 @@ package com.po.fuck.model;
 import com.badlogic.gdx.graphics.g2d.Sprite;
 import com.badlogic.gdx.math.Polygon;
 import com.badlogic.gdx.math.Vector2;
-import com.po.fuck.view.Drawable;
 import com.po.fuck.model.collision.Collidable;
 import com.po.fuck.model.lifetime.Destructable;
