+++ core/src/com/po/fuck/model/weapons/Bullet.java
@@ -3,7 +3,7 @@ package com.po.fuck.model.weapons;
 import com.badlogic.gdx.graphics.g2d.Sprite;
 import com.badlogic.gdx.math.Polygon;
 import com.badlogic.gdx.math.Vector2;
-import com.po.fuck.view.Drawable;
+import com.po.fuck.model.Drawable;
 import com.po.fuck.model.collision.Collidable;
 import com.po.fuck.model.Entity;
 import com.po.fuck.model.GeometryMisc;
